Add url_convert function and create test case
