+++ tests/test_geometries.py
@@ -3,7 +3,6 @@ import numpy.testing as npt
 import pytest
 
 from zndraw import Box, Sphere, ZnDraw
-from zndraw.utils import direction_to_euler, euler_to_direction
 
 
 def test_geometries(server, s22):
@@ -29,24 +28,3 @@ def test_geometries(server, s22):
         vis.geometries = "Geometries are not string!"
 
 
-def test_conversion_utils():
-    """Test conversion functions"""
-    direction = np.array([1, 2, 3])
-    direction = direction / np.linalg.norm(direction)
-    euler = direction_to_euler(direction)
-    new_direction = euler_to_direction(euler)
-
-    npt.assert_allclose(direction, new_direction, atol=1e-6)
-
-    direction = np.array([1, 0, 0])
-    euler = direction_to_euler(direction)
-    new_direction = euler_to_direction(euler)
-
-    npt.assert_allclose(direction, new_direction, atol=1e-6)
-    npt.assert_allclose(euler, [0, 0, 0], atol=1e-6)
-
-    direction = np.array([0, -1, 0])
-    euler = direction_to_euler(direction)
-    new_direction = euler_to_direction(euler)
-
-    npt.assert_allclose(direction, new_direction, atol=1e-6)
